+++ cpp-atm/TransactionDao.h
@@ -14,7 +14,7 @@ private:
 
 	DBController* _db;
 
-	Transaction instance(const string& fromCard, const string& toCard, const float& sum, const string& dateTime);
+	Transaction instance(const string& fromCard, const string& toCard, const double& sum, const string& dateTime);
 
 public:
